Safari-611.1.10/LayoutTests/svg/dynamic-updates/SVGTextElement-dom-x-attr.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGTextElement-dom-x-attr.js
// [Expected rendering result] Text message at 0x20 - and a series of PASS messages

description("Tests dynamic updates of the 'x' attribute of the SVGTextElement object")
createSVGTestCase();

var textElement = createSVGElement("text");
textElement.setAttribute("x", "50");
textElement.setAttribute("y", "20");
textElement.textContent = "Text content";
rootSVGElement.appendChild(textElement);

shouldBeEqualToString("textElement.getAttribute('x')", "50");

function repaintTest() {
    textElement.setAttribute("x", "0");
    shouldBeEqualToString("textElement.getAttribute('x')", "0");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
