third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGTextElement-dom-x-attr-expected.txt

SVG 1.1 dynamic update tests
Text content

Tests dynamic updates of the 'x' attribute of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.getAttribute('x') is "50"
PASS textElement.getAttribute('x') is "0"
PASS successfullyParsed is true

TEST COMPLETE
